quotation update
-updated the quotation email blade
- created a new migration to store messages from contact form
-

+++ database/migrations/2023_03_06_123408_create_items_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->foreign('user_id')->references('user_id')->on('quotations')->onDelete('cascade');
             $table->string('name');
             $table->decimal('price');
-            $table->decimal('qty')->nullable();
+            $table->integer('qty')->nullable();
             $table->decimal('sub_total')->nullable();
             $table->string('QI_id');
             $table->foreign('QI_id')->references('quotation_no')->on('quotations')->onDelete('cascade');
